+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Add `installations` command to explort entities for Giant Swarm installations.
+- Add `installations` command to export entities for Giant Swarm installations.
 
 ## [0.14.0] - 2024-07-10