+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.15.0] - 2024-08-03
+
 ### Added
 
 - Add `installations` command to export entities for Giant Swarm installations.
@@ -245,7 +247,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial code
 
-[Unreleased]: https://github.com/giantswarm/backstage-catalog-importer/compare/v0.14.0...HEAD
+[Unreleased]: https://github.com/giantswarm/backstage-catalog-importer/compare/v0.15.0...HEAD
+[0.15.0]: https://github.com/giantswarm/backstage-catalog-importer/compare/v0.14.0...v0.15.0
 [0.14.0]: https://github.com/giantswarm/backstage-catalog-importer/compare/v0.13.3...v0.14.0
 [0.13.3]: https://github.com/giantswarm/backstage-catalog-importer/compare/v0.13.2...v0.13.3
 [0.13.2]: https://github.com/giantswarm/backstage-catalog-importer/compare/v0.13.1...v0.13.2